+++ .github/workflows/main.yml
@@ -4,6 +4,7 @@ on: [push]
 
 jobs:
   test:
+    name: ${{ matrix.os }}-python${{ matrix.python-version }}
     runs-on: ubuntu-latest
     strategy:
       matrix:
@@ -34,11 +35,11 @@ jobs:
           python -m pip install -r requirements.txt
       - name: Run pytest
         run: |
-          pytest -v --junitxml=test-output${{ matrix.os }}${{ matrix.python-version }}.xml
+          pytest test_ocspchecker.py -v --junitxml=test-output${{ matrix.os }}-python${{ matrix.python-version }}.xml
       - name: Upload test results
         uses: actions/upload-artifact@b4b15b8c7c6ac21ea08fcf65892d2ee8f75cf882 # tag=v4.4.3
         with:
-          name: pytest-results for ${{ matrix.os }}-${{ matrix.python-version }}
+          name: pytest-results for ${{ matrix.os }}-python${{ matrix.python-version }}
           path: "**/test-output.xml"
 #      - name: Upload coverage results
 #        uses: actions/upload-artifact@b4b15b8c7c6ac21ea08fcf65892d2ee8f75cf882 # tag=v4.4.3
